Refactor Artikel component structure; move to new directory and split into ArtikelList and ArtikelCard components for better organization and maintainability
